separate of concerns
