+++ components/link/ArchiveLink.vue
@@ -1,0 +1,19 @@
+<template>
+  <nuxt-link
+    to="/archive"
+    class="btn bg-white border border-solid rounded-sm flex items-center justify-center w-8 h-8 font-bold text-gray p-0 m-0 group hover:text-white hover:bg-primary focus:text-white focus:bg-primary focus:outline-none hover:border hover:border-solid hover:border-primary focus:border focus:border-solid focus:border-primary"
+  >
+    <svg
+      class="svg-inline--fa fa-archive fa-w-16 block w-4 h-4"
+      role="img"
+      viewBox="0 0 512 512"
+    >
+      <title>Archive Icon</title>
+      <path
+        class="fill-primary group-hover:fill-white group-focus:fill-white"
+        d="M32 448c0 17.7 14.3 32 32 32h384c17.7 0 32-14.3 32-32V160H32v288zm160-212c0-6.6 5.4-12 12-12h104c6.6 0 12 5.4 12 12v8c0 6.6-5.4 12-12 12H204c-6.6 0-12-5.4-12-12v-8zM480 32H32C14.3 32 0 46.3 0 64v48c0 8.8 7.2 16 16 16h480c8.8 0 16-7.2 16-16V64c0-17.7-14.3-32-32-32z"
+      />
+    </svg>
+    <span class="sr-only">Archive page</span>
+  </nuxt-link>
+</template>
