+++ pages/index.vue
@@ -10,8 +10,8 @@
 <script lang="ts">
 import Vue from 'vue'
 import axios from 'axios'
-import { TOKEN } from '../static/config'
-import NewsList from '../components/NewsList.vue'
+import { TOKEN } from '@/static/config'
+import NewsList from '@/components/list/NewsList.vue'
 
 interface AsyncData {
   [lists: string]: { data: { api: Array<Object> } }
